+++ spec/nylas/resources/messages_spec.rb
@@ -140,6 +140,8 @@ describe Nylas::Messages do
     it "calls the post method with the correct parameters and attachments" do
       identifier = "abc-123-grant-id"
       mock_file = instance_double("file")
+      allow(mock_file).to receive(:read).and_return("file content")
+      allow(mock_file).to receive(:close).and_return(true)
       request_body = {
         subject: "Hello from Nylas!",
         to: [{ name: "Jon Snow", email: "jsnow@gmail.com" }],